feat: deploy + initialize on devnet, fix Anchor 0.31 usize bug
Program HkbccHJ45V7zbgLkwr64EUzRhfjdH1mcoQ5UVMAte341 deployed and
GlobalConfig PDA initialized on Solana devnet.

- migrations/deploy.ts: idempotent initialize() runner; supports
  both `anchor migrate` (passes provider) and direct `npx ts-node`
  execution via ANCHOR_WALLET + ANCHOR_PROVIDER_URL env vars.
- package.json: bump @coral-xyz/anchor to 0.31.1 to match
  Anchor.toml + anchor-cli, add @solana/web3.js and ts-node.
- constants.rs: change MAX_*_LEN from `usize` to `u64`. Anchor
  0.31's IDL builder expands `#[constant]` into a call to
  `usize::get_full_path()` which doesn't exist; fixed-width
  integers work fine.
- register_authority.rs / mint_vehicle_passport.rs: cast
  `String::len()` (usize) to u64 when comparing against the
  new u64 constants.
- app/lib/idl.json + program-types.ts: exported from
  `target/idl/` and `target/types/` so the Next.js client can
  build an Anchor `Program<VehicleHistory>` without depending
  on the Rust build artifact tree.

+++ programs/vehicle_history/src/constants.rs
@@ -12,11 +12,14 @@ pub const VEHICLE_SEED: &[u8] = b"vehicle";
 #[constant]
 pub const EVENT_SEED: &[u8] = b"event";
 
+// IDL builder in Anchor 0.31 doesn't accept `usize` for `#[constant]` (calls
+// `usize::get_full_path()` which doesn't exist). Using `u64` and casting at
+// the call site keeps the constants exposed in the IDL.
 #[constant]
-pub const MAX_NAME_LEN: usize = 64;
+pub const MAX_NAME_LEN: u64 = 64;
 
 #[constant]
-pub const MAX_MAKE_LEN: usize = 32;
+pub const MAX_MAKE_LEN: u64 = 32;
 
 #[constant]
-pub const MAX_MODEL_LEN: usize = 32;
+pub const MAX_MODEL_LEN: u64 = 32;
